+++ crates/algod_client/src/models/avm_value.rs
@@ -23,7 +23,7 @@ pub struct AvmValue {
     pub bytes: Option<String>,
     /// uint value.
     #[serde(rename = "uint", skip_serializing_if = "Option::is_none")]
-    pub uint: Option<i32>,
+    pub uint: Option<u64>,
 }
 
 impl AlgorandMsgpack for AvmValue {
