+++ crates/algod_client/src/models/transaction_params200_response.rs
@@ -27,7 +27,7 @@ pub struct TransactionParams200Response {
     /// Fee may fall to zero but transactions must still have a fee of
     /// at least MinTxnFee for the current network protocol.
     #[serde(rename = "fee")]
-    pub fee: i32,
+    pub fee: u64,
     /// GenesisHash is the hash of the genesis block.
     #[serde_as(as = "serde_with::base64::Base64")]
     #[serde(rename = "genesis-hash")]
@@ -37,11 +37,11 @@ pub struct TransactionParams200Response {
     pub genesis_id: String,
     /// LastRound indicates the last round seen
     #[serde(rename = "last-round")]
-    pub last_round: i32,
+    pub last_round: u64,
     /// The minimum transaction fee (not per byte) required for the
     /// txn to validate for the current network protocol.
     #[serde(rename = "min-fee")]
-    pub min_fee: i32,
+    pub min_fee: u64,
 }
 
 impl AlgorandMsgpack for TransactionParams200Response {
@@ -52,11 +52,11 @@ impl TransactionParams200Response {
     /// Constructor for TransactionParams200Response
     pub fn new(
         consensus_version: String,
-        fee: i32,
+        fee: u64,
         genesis_hash: Vec<u8>,
         genesis_id: String,
-        last_round: i32,
-        min_fee: i32,
+        last_round: u64,
+        min_fee: u64,
     ) -> TransactionParams200Response {
         TransactionParams200Response {
             consensus_version,
